backend/requirements.txt
# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9

# AI Providers — model-agnostic, swap freely
anthropic==0.34.0
openai==1.51.0
google-generativeai==0.8.0

# Voice / Speech
openai-whisper==20240930

# Memory & Embeddings
pgvector==0.3.2
sentence-transformers==3.1.1

# Caching
redis==5.1.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
tenacity==9.0.0
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
